Move towards a reusable selection field component
